gencart_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow>=10.0.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==23.4

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0

# Packaging (needed for pkg_resources and building wheels)
setuptools>=70.0.0
wheel>=0.43.0

# Basic ML Dependencies
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8
textblob>=0.17.0
joblib>=1.3.0

# Optional Advanced ML (install separately if needed)
# transformers>=4.30.0
# torch>=2.0.0
# tensorflow>=2.13.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
